Fix issue adding blockquote with alternate data (#11)
This PR fixes an issue where selecting text to quote doesn't work if
there is other text already in the comment.

Fixes: https://github.com/thunder-app/thunder/issues/1726

Will need a corresponding change in Thunder to update the package.

Here I demo that the problem is fixed, as well as a regression test on
other types of block quotes.


https://github.com/user-attachments/assets/366b9c86-4e3c-4e31-b53f-cb8cd6c2458b